==> code/3_Split_data_train_val_test.ipynb <==
import numpy as np
import pandas as pd
import sys
import os
import matplotlib.pyplot as plt
%matplotlib inline
import scipy
import h5py
import tables
import glob
import tables
import tqdm
import re

np.random.seed(10)
basepath = os.path.dirname(os.getcwd()) + "/"
hasepath = basepath + "/hase/"
savepath = basepath + "/processed_data/"
studyname = str(np.load(savepath + "studyname.npy"))

# Split Dataset

This is merely an example since there is a wide variety in how phenotype is registered. We assume here an csv file with the same patient ordering as in the genotype file.

df_pheno = pd.read_csv(datapath + "phenotype_file.csv")

# if there is the same ordering as in the genotype matrix the following can be used:

df_pheno["tot_index"] = np.arange(len(df_pheno))

# otherwise make sure that there exists a column 'tot_index' with the correct indices for the genotype file.

print("phenotype all",df_pheno.shape)
df_pheno.head()

phenotype_name = "Sex"

if phenotype_name == "Smoking" :
    
    print("never smokers = ",sum(df_pheno["Smoking"] == 0))
    print("former smokers = ",sum(df_pheno["Smoking"] == 1))
    print("current smokers = ",sum(df_pheno["Smoking"] == 2))
    df_pheno["labels"] =(df_pheno["Smoking"]>0)*1 
    print("ever smoked =", sum(df_pheno["labels"]))
    
if phenotype_name == "Sex":
    print("male = ",sum(df_pheno["Sex"] == 0))
    print("female = ",sum(df_pheno["Sex"] == 1))
    print(df_pheno["Sex"].unique())
    df_pheno["labels"] =(df_pheno["Smoking"]>0)*1 

#  Split cases
df_case = df_pheno[df_pheno["labels"] == 1]
df_pheno_train = df_case.sample(frac = 0.6, random_state=0)
restdata = df_case.loc[~df_case.index.isin(df_pheno_train.index)]
df_pheno_val = restdata.sample(frac=0.5, random_state=0)
df_pheno_test = restdata.loc[~restdata.index.isin(df_pheno_val.index)]

#  Split controls
df_control = df_pheno[(df_pheno["labels"] == 0) ]
df_control_train = df_control.sample(frac = 0.6, random_state=0)
restdata = df_control.loc[~df_control.index.isin(df_control_train.index)]
df_control_val = restdata.sample(frac =0.5, random_state=0)
df_control_test = restdata.loc[~restdata.index.isin(df_control_val.index)]

ytrain_pheno = pd.concat([df_pheno_train, df_control_train])
yval_pheno = pd.concat([df_pheno_val, df_control_val])
ytest_pheno = pd.concat([df_pheno_test, df_control_test])

ytrain_pheno.to_csv(datapath +  "/ytrain_"+phenotype_name+".csv")
yval_pheno.to_csv(datapath + "/yval_"+phenotype_name+".csv")
ytest_pheno.to_csv(datapath + "/ytest_"+phenotype_name+".csv")

print("total = ",ytrain_pheno.shape)
print("class 1 = ",ytrain_pheno["labels"].sum())
print("class 0 = ",sum(ytrain_pheno["labels"] == 0))
print("ratio = ", 1/ (ytrain_pheno["labels"].sum()/ ytrain_pheno.shape[0]) )

print("total = ",yval_pheno.shape)
print("class 1 = ",yval_pheno["labels"].sum())
print("class 0 = ",sum(yval_pheno["labels"]==0))
print("ratio = ", 1/(yval_pheno["labels"].sum() / yval_pheno.shape[0]) )

print("total = ",ytest_pheno.shape)
print("class 1 = ", ytest_pheno["labels"].sum())
print("class 0 = ",sum(ytest_pheno["labels"]==0))
print("ratio = ",1/(ytest_pheno["labels"].sum() /  ytest_pheno.shape[0] ) )
